requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
docker-pycreds==0.4.0
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.28.1
gym==0.17.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rl==0.4.2
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pid==3.0.3
Pillow==7.1.2
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
wandb==0.8.35
watchdog==0.10.2
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
